[linked editing] finalize LinkedEditingRangeProvider. Fixes #109923
